### 2022-05-04 #4375 @ThomasThelen
Potential reviewers:
brunomontezano
babakmohammadi
thodson-usgs
MoLi7
chrisaberson
