### 2022-05-18 baf7c83
Remove the URL prefix from the DOI
