+++ paper/paper.bib
@@ -47,7 +47,7 @@
 	volume       = 25,
 	number       = 8,
 	pages        = {972--974},
-	doi          = {https://doi.org/10.1016/j.envsoft.2010.02.008},
+	doi          = {10.1016/j.envsoft.2010.02.008},
 	issn         = {1364-8152},
 	url          = {https://www.sciencedirect.com/science/article/pii/S1364815210000514},
 	keywords     = {NetLogo, Gnu R, Model analysis, Agent-based modelling, Individual-based modelling, Modelling software},
