### 2022-05-18 #4375 @ThomasThelen
Hi @marcosvital,

Thanks for volunteering to be the editor for this submission. I fixed the DOI issue in [baf7c83](https://github.com/zizroc/villager/commit/baf7c83b6f5144231441e95d64077661d2b17fe9) and it should no longer be an issue. I've gone back and reviewed the submission and would like to continue with it.
