### 2022-06-26 #4375 @chrisaberson
Agree to review

On Fri, Jun 24, 2022, 3:38 PM Marcos Vital (LEQ-UFAL) <
***@***.***> wrote:

> Dear @brunomontezano <https://github.com/brunomontezano>, @chrisaberson
> <https://github.com/chrisaberson> and @MoLi7 <https://github.com/MoLi7>
> you be willing to review this submission for JOSS?
>
> We carry out an open checklist-driven reviews here in GitHub issues, and
> follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> This is still a pre-review issue, and once there are enough reviewers we
> will move on to a new issue where the review will take place. If you are
> not able to review, please let us know if you have recommendations for
> other reviewers that could help with this submission.
>
