### 2022-07-06 #4375 @thodson-usgs
I have no experience with agent-based modeling, but I would be willing to provide a checklist-driven-common-sense review.
