### 2022-07-11 #4375 @marcosvital
> I have no experience with agent-based modeling, but I would be willing to provide a checklist-driven-common-sense review.
> […](#)

Hi, @thodson-usgs! Since we will have three different reviewers for this submission, it's fine if you have no experience with this theme. I'll add you as a reviewer.
