### 2022-07-11 #4562 @editorialbot
[REVIEW]: villager: A framework for designing and executing agent-based models in R
**Submitting author:** @thomasthelen (<a href="http://orcid.org/0000-0002-1756-2128">Thomas Thelen</a>)
**Repository:** https://github.com/zizroc/villager
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.2.1
