### 2022-07-11 #4562 @marcosvital
Dear @chrisaberson, @brunomontezano and @thodson-usgs, thank you again for accepting review this submission for JOSS. The reviewing process is checklist based, and instructions were already posted above by the editorial bot - but let me know if you need any assistance, ok? Also, you can tag @ThomasThelen if you have specific questions about the manuscript.

@ThomasThelen, you can tag any of your co-authors if you want, so they would be able to follow this issue if they use GitHub.
