### 2022-07-11 4d4bb31
Update paper.md
1. Minor revisions
1. Removed dependencies section. This should be included in the documentation, but I don't think it's necessary for the paper.
1. Added Acknowledgements section. Technically not required, but I suggest you at least acknowledge any funding support, for example https://github.com/chasmani/piecewise-regression/blob/master/paper/paper.md
