### 2022-07-11 zizroc/villager#85 @thodson-usgs
Update license for JOSS 
openjournals/joss-reviews/issues/4562:

1. Delete either `LICENSE' of `LICENSE.md`
2. Use a standard template for `LICENSE`.
3. Update copyright holder information. I believe this should be one of the authors or their institution (For example: https://github.com/chasmani/piecewise-regression/blob/master/LICENSE.txt). Joint ownership is also permitted.
