### 2022-07-12 #4562 @chrisaberson
I would like to see a clearer, real-world example. I recognize that this is a complex method but the present example really only will be of use to experts
