### 2022-07-12 #4562 @ThomasThelen
That can definitely be done. I can take a look at more classical motivating ABM examples and create an additional vignette demonstrating it under this framework.
