### 2022-07-19 #4562 @thodson-usgs
Technically a real-world example is not required in our reviewer checklist:
"Example usage: Do the authors include examples of how to use the software (*ideally* to solve real-world analysis problems)."
I think the authors have minimally satisfied that requirement, but @chrisaberson is right.
