### 2022-07-23 1b893f0
Fix gramatical issues pointed out by thodson-usgs
