+++ paper/paper.bib
@@ -1,19 +1,19 @@
 @manual{urbanek_2020,
-	title        = {uuid: Tools for Generating and Handling of UUIDs},
+	title        = {{uuid}: Tools for Generating and Handling of {UUID}s},
 	author       = {Simon Urbanek and Theodore Ts'o},
 	year         = 2020,
 	url          = {https://CRAN.R-project.org/package=uuid},
 	note         = {R package version 0.1-4}
 }
 @manual{chang_2020,
-	title        = {R6: Encapsulated Classes with Reference Semantics},
+	title        = {{R6}: Encapsulated Classes with Reference Semantics},
 	author       = {Winston Chang},
 	year         = 2020,
 	url          = {https://CRAN.R-project.org/package=R6},
 	note         = {R package version 2.5.0}
 }
 @manual{wickam_2020,
-	title        = {readr: Read Rectangular Text Data},
+	title        = {{readr}: Read Rectangular Text Data},
 	author       = {Hadley Wickham and Jim Hester},
 	year         = 2020,
 	url          = {https://CRAN.R-project.org/package=readr},
@@ -40,7 +40,7 @@
 	version      = {v2.0.8}
 }
 @article{THIELE2010972,
-	title        = {NetLogo meets R: Linking agent-based models with a toolbox for their analysis},
+	title        = {{NetLogo} meets {R}: Linking agent-based models with a toolbox for their analysis},
 	author       = {Jan C. Thiele and Volker Grimm},
 	year         = 2010,
 	journal      = {Environmental Modelling & Software},
