+++ tests/testthat/test-resource.R
@@ -5,7 +5,7 @@ test_that("the default constructor works", {
   res_quantity <- 11
   test_resource <- resource$new(name = res_name, quantity = res_quantity)
 
-  # Make sure that new winik's start at age 0
+  # Make sure that new agent's start at age 0
   testthat::expect_equal(test_resource$name, res_name)
 
   # Also make sure that they're alive
