### 2022-08-07 36c387a
Update paper to remove references to 'winik'
