+++ man/Simulation.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/Simulation.R
+% Please edit documentation in R/simulation.R
 \docType{class}
 \name{simulation}
 \alias{simulation}
