### 2022-08-07 #4562 @ThomasThelen
Sorry for the delay folks, but I should have all of the feedback thus far addressed. I cut a new release, 1.2.0 which includes all of the great feedback.

@thodson-usgs: I addressed all of your feedback about the license, wording in the paper, and convention of using (agent). Thank you!
@chrisaberson: I added a [new vignette](https://github.com/zizroc/villager/blob/main/vignettes/example-boltzman-wealth.Rmd) that resembles an example from Project Mesa (the canonical Python ABM framework)
With those two addressed, we should be down to one reviewer left.
@brunomontezano: Let me know if there's anything I can do to help facilitate the review
