+++ README.md
@@ -66,14 +66,14 @@ test_model <- function(current_state, previous_state, model_data, agent_mgr, res
 }
 ```
 
-The agent manager can also be used to pair agents, representitive of a relationship or social bond.
+The agent manager can also be used to pair agents, representative of a relationship or social bond.
 ```
 agent_mgr$agent_mgr$connect_agents(mother, father)
 ```
 
 ### Creating & Managing Resources
 
-Resources are similar to agents in that they're both R6 classes, are instantiated similarly, and are also managed by an object passed into the model. An example of creating resources and adding them to the simualtion is given below.
+Resources are similar to agents in that they're both R6 classes, are instantiated similarly, and are also managed by an object passed into the model. An example of creating resources and adding them to the simulation is given below.
 
 ```
 test_model <- function(current_state, previous_state, model_data, agent_mgr, resource_mgr) {
@@ -90,7 +90,7 @@ test_model <- function(current_state, previous_state, model_data, agent_mgr, res
 
 ### State
 
-Objects of type `village`, `agent`, and `resource`have particular states at a particular time. As the simulation progresses, the state of these change based on model logic. At the end of each time step, the state of each object is saved, giving a complete record of the system's evolution. The essence of any agent based model is changing the state at each time step. villager provides a mechanim for defining the initial state and for changing the state throughout the simulation.
+Objects of type `village`, `agent`, and `resource`have particular states at a particular time. As the simulation progresses, the state of these change based on model logic. At the end of each time step, the state of each object is saved, giving a complete record of the system's evolution. The essence of any agent based model is changing the state at each time step. villager provides a mechanism for defining the initial state and for changing the state throughout the simulation.
 
 ### Managing the Initial State
 
