+++ man/data_writer.Rd
@@ -9,7 +9,7 @@ A class responsible for the simulation data to disk.
 }
 \details{
 This class can be subclasses to provide advanced data writing to other data sources. This should also
-be subclassed if the agent and resource classes are subclasses, to write any addional fields to the data source.
+be subclassed if the agent and resource classes are subclasses, to write any additional fields to the data source.
 }
 \section{Methods}{
 
