+++ man/model_data.Rd
@@ -2,14 +2,9 @@
 % Please edit documentation in R/model_data.R
 \name{model_data}
 \alias{model_data}
-\title{R6 Class representing data that's external from resources and agents}
+\title{Model Data}
 \description{
 R6 Class representing data that's external from resources and agents
-
-R6 Class representing data that's external from resources and agents
-}
-\details{
-It contains a single variable, 'events' for when the data holds a list of events
 }
 \section{Public fields}{
 \if{html}{\out{<div class="r6-fields">}}
