+++ NAMESPACE
@@ -3,8 +3,12 @@
 export(agent)
 export(agent_manager)
 export(data_writer)
+export(model_data)
 export(resource)
 export(resource_manager)
 export(simulation)
 export(village)
 export(village_state)
+importFrom(R6,R6Class)
+importFrom(readr,write_csv)
+importFrom(uuid,UUIDgenerate)
