+++ R/model_data.R
@@ -1,7 +1,7 @@
 #' @export
 #' @title Model Data
 #' @description R6 Class representing data that's external from resources and agents
-#' @docType class
+#'
 #' It contains a single variable, 'events' for when the data holds a list of events
 model_data <- R6::R6Class("model_data",
                           public = list(
