### 2022-09-25 3f2a8ab
Fix formatting of population example
