### 2022-10-25 #4562 @kthyng
Hi @ThomasThelen! Please update the metadata on your zenodo archive so that the title and author list exactly match your JOSS paper.
