### 2022-10-25 #4562 @kthyng
Edits for your paper @ThomasThelen:
* The citation for McIntire on the first page isn't placed properly — please fix up the syntax for parenthetical citing.
* the citations in your "Dependencies" section should be inline instead of parenthetical
You can see an example [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)
