### 2022-11-05 #4562 @ThomasThelen
Hey @kthyng sorry for the delay! Thanks for the catch with those citations-they should be fixed now. I moved the parenthetical ones inline with the first column so that they're no longer detached. I'll take a look at the Zenodo archive now.
