### 2022-11-07 #4562 @kthyng
@ThomasThelen Ok great, your paper looks good now. let me know when you've been able to work on the Zenodo metadata.
