### 2022-11-10 #4562 @kthyng
Congrats on your new publication @ThomasThelen! Many thanks to editor @marcosvital and reviewers @chrisaberson, @brunomontezano, and @thodson-usgs for your time, hard work, and expertise!
